requirements.txt
langchain
langchain-core
langchain-community
langchain-anthropic
langchain-google-genai
google-generativeai
langchain-huggingface
transformers
huggingface-hub
sentence-transformers
python-dotenv
numpy
scikit-learn
torch
torchvision
torchaudio
streamlit
pypdf
pymupdf
faiss-cpu
wikipedia
duckduckgo-search
arxiv